Revert "video: tegra: host: gk20a: Do not suspend if active"
This reverts commit 63078e87a62b265111622e58e66d9e67a2e1519b. The
original patch was added to protect kernel from applications that
injected jobs just after pm had suspended the device. This lead
the bookkeeping to get mixed as some callbacks were not called
correctly before the device was suspended.

We have now revisited the suspend sequences and identified that
we actually used to use some callbacks in a wrong way and the
original patch was simply a workaround to hide the issue. Now
that the issues are fixed, we can safely remove the hack.

Bug 1426717

Change-Id: I42d884bc21baad1a37001c354ef1a80e2eba140c
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/358575
(cherry picked from commit 12d0c2928fb62b4ba649dcd88b6fa0b057f1182c)
Reviewed-on: http://git-master/r/335100
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>